requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
asyncpg==0.29.0
python-multipart==0.0.6
aiofiles==23.2.1